requirements.txt
pyserial==3.5
opencv-python==4.9.0.80
Pillow==11.2.1
ultralytics==8.3.158
numpy==1.26.4
matplotlib==3.8.4
reportlab
